requirements.txt
langchain>=0.1.0
streamlit>=1.25.0
chromadb>=0.4.0
huggingface_hub>=0.17.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
requests>=2.31.0 